backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
redis>=5.2.0
celery>=5.4.0
httpx>=0.28.0
python-multipart>=0.0.17
python-jose[cryptography]>=3.3.0
anthropic>=0.39.0
google-generativeai>=0.8.0
aiofiles>=24.1.0
